requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==2.2.10
django-cleanup==5.1.0
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-cron==0.5.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
idna==2.10
Pillow==8.0.1
psycopg2
PyJWT==1.7.1
pytz==2020.1
requests==2.25.0
sqlparse
stripe==2.55.1
urllib3==1.26.1
wincertstore==0.2
